Chore: completed all MVPs
